+++ demo/hello-world/run.py
@@ -62,7 +62,7 @@ def main():
             "description": "",
             "config": {
                 "transport": {
-                    "name": "file",
+                    "name": "file_" + ("input" if is_input else "output"),
                     "config": {
                         "path": filepath,
                     }